LayoutTests/inspector/canvas/recording-bitmaprenderer-memoryLimit-expected.txt

Test that CanvasManager is able to record actions made to BitmapRenderer canvas contexts.


== Running test suite: Canvas.recordingBitmapRenderer
-- Running test case: Canvas.recordingBitmapRenderer.memoryLimit
initialState:
  attributes:
    width: 2
    height: 2
  parameters:
  content: <filtered>
frames:
  0: (duration) (incomplete)
    0: transferFromImageBitmap([object ImageBitmap])
      swizzleTypes: [ImageBitmap]
      trace:
        0: transferFromImageBitmap
        1: (anonymous function)
        2: executeFrameFunction
        3: (anonymous function)
        4: asyncFunctionResume
        5: (anonymous function)
        6: promiseReactionJobWithoutPromise
      snapshot: <filtered>
